Move list logic to new component RestaurantList
